backend/requirements.txt
django>=5.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.4
mysqlclient>=2.2
python-decouple>=3.8
Pillow>=10.0
gunicorn>=22.0
